Supported enabling/disabling the feature of spawning new processes for training.

+++ plato/servers/base.py
@@ -79,7 +79,7 @@ class Server:
     def run(self, client=None, edge_server=None, edge_client=None):
         """Start a run loop for the server. """
         # Remove the running trainers table from previous runs.
-        if not Config().is_edge_server():
+        if not Config().is_edge_server() and hasattr(Config().trainer, 'max_concurrency'):
             with Config().sql_connection:
                 Config().cursor.execute("DROP TABLE IF EXISTS trainers")
