+++ plato/trainers/basic.py
@@ -6,7 +6,6 @@ import logging
 import multiprocessing as mp
 import os
 import time
-from typing import Tuple
 
 import numpy as np
 import torch
@@ -202,14 +201,15 @@ class Trainer(base.Trainer):
 
         self.model.cpu()
 
-        model_type = config['model_name']
-        filename = f"{model_type}_{self.client_id}_{config['run_id']}.pth"
-        self.save_model(filename)
+        if 'max_concurrency' in config:
+            model_type = config['model_name']
+            filename = f"{model_type}_{self.client_id}_{config['run_id']}.pth"
+            self.save_model(filename)
 
         if 'use_wandb' in config:
             run.finish()
 
-    def train(self, trainset, sampler, cut_layer=None) -> Tuple[bool, float]:
+    def train(self, trainset, sampler, cut_layer=None) -> float:
         """The main training loop in a federated learning workload.
 
         Arguments:
@@ -219,44 +219,50 @@ class Trainer(base.Trainer):
 
         Returns:
         bool: Whether training was successfully completed.
-        float: The training time.
+        float: Elapsed time during training.
         """
-        self.start_training()
-        tic = time.perf_counter()
-
-        if mp.get_start_method(allow_none=True) != 'spawn':
-            mp.set_start_method('spawn', force=True)
-
         config = Config().trainer._asdict()
         config['run_id'] = Config().params['run_id']
 
-        train_proc = mp.Process(target=Trainer.train_process,
-                                args=(
-                                    self,
-                                    config,
-                                    trainset,
-                                    sampler,
-                                    cut_layer,
-                                ))
-        train_proc.start()
-        train_proc.join()
+        if hasattr(Config().trainer, 'max_concurrency'):
+            self.start_training()
+            tic = time.perf_counter()
 
-        model_name = Config().trainer.model_name
-        filename = f"{model_name}_{self.client_id}_{Config().params['run_id']}.pth"
-        try:
-            self.load_model(filename)
-        except OSError:  # the model file is not found, training failed
-            logging.info("The training process on client #%d failed.",
-                         self.client_id)
-            self.run_sql_statement("DELETE FROM trainers WHERE run_id = (?)",
-                                   (self.client_id, ))
-            return False, 0.0
+            if mp.get_start_method(allow_none=True) != 'spawn':
+                mp.set_start_method('spawn', force=True)
 
-        toc = time.perf_counter()
-        self.pause_training()
+            train_proc = mp.Process(target=Trainer.train_process,
+                                    args=(
+                                        self,
+                                        config,
+                                        trainset,
+                                        sampler,
+                                        cut_layer,
+                                    ))
+            train_proc.start()
+            train_proc.join()
+
+            model_name = Config().trainer.model_name
+            filename = f"{model_name}_{self.client_id}_{Config().params['run_id']}.pth"
+
+            try:
+                self.load_model(filename)
+            except OSError as error:  # the model file is not found, training failed
+                if hasattr(Config().trainer, 'max_concurrency'):
+                    self.run_sql_statement("DELETE FROM trainers WHERE run_id = (?)",
+                                        (self.client_id, ))
+                raise ValueError(f"Training on client {self.client_id} failed.") from error
+
+            toc = time.perf_counter()
+            self.pause_training()
+        else:
+            tic = time.perf_counter()
+            self.train_process(config, trainset, sampler, cut_layer)
+            toc = time.perf_counter()
+
         training_time = toc - tic
 
-        return True, training_time
+        return training_time
 
     def test_process(self, config, testset):
         """The testing loop, run in a separate process with a new CUDA context,
@@ -264,7 +270,6 @@ class Trainer(base.Trainer):
 
         Arguments:
         config: a dictionary of configuration parameters.
-        rank: Required by torch.multiprocessing to spawn processes. Unused.
         testset: The test dataset.
         """
         self.model.to(self.device)
@@ -300,9 +305,12 @@ class Trainer(base.Trainer):
 
         self.model.cpu()
 
-        model_name = config['model_name']
-        filename = f"{model_name}_{self.client_id}_{config['run_id']}.acc"
-        Trainer.save_accuracy(accuracy, filename)
+        if 'max_concurrency' in config:
+            model_name = config['model_name']
+            filename = f"{model_name}_{self.client_id}_{config['run_id']}.acc"
+            Trainer.save_accuracy(accuracy, filename)
+        else:
+            return accuracy
 
     def test(self, testset) -> float:
         """Testing the model using the provided test dataset.
@@ -313,30 +321,32 @@ class Trainer(base.Trainer):
         config = Config().trainer._asdict()
         config['run_id'] = Config().params['run_id']
 
-        self.start_training()
+        if hasattr(Config().trainer, 'max_concurrency'):
+            self.start_training()
 
-        if mp.get_start_method(allow_none=True) != 'spawn':
-            mp.set_start_method('spawn', force=True)
+            if mp.get_start_method(allow_none=True) != 'spawn':
+                mp.set_start_method('spawn', force=True)
 
-        proc = mp.Process(target=Trainer.test_process,
-                          args=(
-                              self,
-                              config,
-                              testset,
-                          ))
-        proc.start()
-        proc.join()
+            proc = mp.Process(target=Trainer.test_process,
+                            args=(
+                                self,
+                                config,
+                                testset,
+                            ))
+            proc.start()
+            proc.join()
 
-        try:
-            model_name = Config().trainer.model_name
-            filename = f"{model_name}_{self.client_id}_{Config().params['run_id']}.acc"
-            accuracy = Trainer.load_accuracy(filename)
-        except OSError:  # the model file is not found, training failed
-            logging.info("The testing process on client #%d failed.",
-                         self.client_id)
-            return 0
+            try:
+                model_name = Config().trainer.model_name
+                filename = f"{model_name}_{self.client_id}_{Config().params['run_id']}.acc"
+                accuracy = Trainer.load_accuracy(filename)
+            except OSError as error:  # the model file is not found, training failed
+                raise ValueError(f"Testing on client #{self.client_id} failed.") from error
 
-        self.pause_training()
+            self.pause_training()
+        else:
+            accuracy = self.test_process(config, testset)
+
         return accuracy
 
     async def server_test(self, testset):
